// webpack---_N_E/libs/js/sapi/src/types.ts
export type CreatorMetadata = {
  appearanceCustomization: AppearanceCustomizationState
  avatarData: AvatarData
  avatarImageURL: string
  displayName: string
  id: string
  username: string
}

type SpaceTrailer = {
  poster: string
  video: string
}

export type SpaceMetadata = {
  active?: boolean
  activeUserCount: number
  customVideo?: SpaceTrailer
  description?: string
  id: string
  joinCount: number
  likeCount: number
  liked: boolean
  live?: boolean
  lobbyType: SAPILobbyType
  name: string
  ownerID: string
  published: boolean
  roomAdmins: string[]
  /** All thumbnails available for a space **/
  roomThumbnails: SAPIRoomThumbnail
  /** True if the space is the Creator Toolkit sandbox **/
  sandbox?: boolean
  shareID?: string
  slug: string
  tags?: string[]
  /** The thumbnail to use when previewing a space (i.e. in the spaces list) **/
  thumbnail: string
  /** Present on Creator Toolkit spaces */
  worldID?: string
}

export type SpaceAndCreator = {
  creator: CreatorMetadata
  space: SpaceMetadata
}
// ...
export interface SAPIRoomThumbnail {
  cubemapGetUrl: string
  customGetUrl: string
  expiresAt: number
  persistentGetUrl: string
  transientGetUrl: string
}

export interface SAPIAppearanceCustomization {
  profileColor: string
}

type UserAvatarStyle = "CREATOR_TOOLKIT" | "READY_PLAYER_ME" | "REALISTIC_RPM_HYBRID" | "REALISTIC"

export interface ReadyPlayerMeThumbnails {
  lowerBody?: string
  upperBody?: string
}

export interface AvatarData {
  activeAvatarStyle?: UserAvatarStyle
  avatarBody?: string
  avatarID?: string
  avatarUserID?: string
  filename?: string
  readyPlayerMeThumbnails?: ReadyPlayerMeThumbnails
  readyPlayerMeUrl?: string
  shirtColorOverride?: string
  skinColor?: string
  skinColorOverride?: string
}
// ...
export enum SAPILobbyType {
  None = "",
  Org = "ORG_LOBBY",
  Private = "PRIVATE_LOBBY",
  Public = "PUBLIC_LOBBY",
}
// ...
type BadgeBase = {
  badgeIconURL: string
  createdAt: string
  description: string
  id: string
  name: string
  /** ex: 2023-03-04T00:04:57.704328Z */
  rewardedAt?: string
  updatedAt: string
  worldID: string
  worldName: string
}

export type Badge =
  | BadgeBase
  | (BadgeBase & {
      shareID: string
      spaceID: string
      spaceName: string
    })

export type GetBadgeRequest = {
  badgeId: string
}

export type GetBadgeResponse = Badge

export type ClaimBadgeRequest = {
  badgeId: string
  spaceID: string
}

export type UserBadge =
  | BadgeBase
  | (BadgeBase & {
      shareID: string
      spaceID: string
      spaceName: string
    })

export enum FeedComponentType {
  SpacesGrid = "spaces-grid",
  SpacesRow = "spaces-row",
  SpacesTrailers = "spaces-trailers",
  SpacesVideoHighlights = "spaces-video-highlights",
}

/**
 * All possible data types for a feed component.
 */
export type FeedComponentData = SpacesGridFeedComponentData &
  SpacesRowFeedComponentData &
  SpacesTrailersFeedComponentData

export type FeedComponent = {
  data: FeedComponentData
  type: FeedComponentType
}

type FeedTitleSegment = {
  label: string
  link?: string
}

export type FeedTitle = FeedTitleSegment[]
// ...
export type SpacesGridFeedComponentData = {
  analyticsContext: AnalyticsContext
  id: string
  /** Used for "See More" link */
  link: string
  numRows?: 1 | 2 | 3
  spaces: SpaceAndCreator[]
  title: FeedTitle
}

export type SpacesRowFeedComponentData = {
  analyticsContext: AnalyticsContext
  id: string
  link: string
  spaces: SpaceAndCreator[]
  title: FeedTitle
}

export type SpacesTrailersFeedComponentData = {
  analyticsContext: AnalyticsContext
  id: string
  spaces: FeedVideo[]
  title: FeedTitle
}

export type FeedSection = {
  id: string
  numRows?: number
  overrideId?: string
  title: FeedTitle
}

type FeedConfigAuthor = {
  id: string
  playerColor: string
  profilePicUrl: string
  username: string
}

export type FeedConfig = {
  config: FeedSection[]
  createdAt: string
  createdBy?: FeedConfigAuthor
  description: string
}

export type FeedCategory = {
  createdAt?: string
  isPrivate: boolean
  name: string
  slug: string
  spaces: string[]
  tags: string[]
  updatedAt?: string
}